fix(notification): extract deduplication logic to shared utility
Extract notification deduplication logic from NotificationBell into a reusable
utility module and apply it to the Notifications view to fix duplicate messages.

## Changes

- **New**: `src/lib/utils/notifications.ts` - Centralized deduplication utilities
- **Updated**: `NotificationBell.svelte` - Uses shared deduplication logic (-150 LOC)
- **Updated**: `Notifications.svelte` - Added deduplication to API calls

## Fixes

- Resolves duplicate diskmanager error notifications in Notifications view
- Maintains consistent deduplication behavior across components
- Single source of truth for notification logic
